+++ xedocs/schemas/corrections/base_references.py
@@ -92,3 +92,19 @@ class BaseMap(BaseResourceReference):
     algorithm: Literal["cnn", "gcn", "mlp"] = rframe.Index()
 
     value: str
+
+    @property
+    def fmt(self):
+        for fmt in self._ALLOWED_FORMATS:
+            if self.value.endswith(fmt):
+                break
+        else:
+            raise ValueError(f"Unknown format for {self.value}")
+        return fmt
+
+    @property
+    def map(self):
+        import straxen
+
+        map_file = straxen.get_resource(self.value, fmt=self.fmt)
+        return straxen.InterpolatingMap(map_file)
